Fix env_universal_common.cpp filesize comparisons.
* use off_t instead of size_t for file size where it can mismatch
  st_size's type in stat.h (cherry-pick merge omitted this)

+++ src/env_universal_common.cpp
@@ -1175,9 +1175,9 @@ class universal_notifier_shmem_poller_t : public universal_notifier_t
             report_error(err, L"Unable to open shared memory with path '%s'", path);
             errored = true;
         }
-
+		
         /* Get the size */
-        size_t size = 0;
+        off_t size = 0;
         if (! errored)
         {
             struct stat buf = {};
